requirements.txt
altgraph==0.17
astroid==2.3.3
isort==4.3.21
lazy-object-proxy==1.4.3
macholib==1.14
mccabe==0.6.1
numpy==1.18.1
Pillow==7.0.0
pylint==2.4.4
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
typed-ast==1.4.1
wrapt==1.11.2
